Move Logger.Level formatting into its own file
